requirements.txt
# Google API Libraries
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0

# Google Gemini AI
google-generativeai==0.3.2

# Data Processing
pandas==2.1.4
numpy==1.26.2
openpyxl==3.1.2

# PDF Processing
PyPDF2==3.0.1
pdfplumber==0.10.3

# Environment & Configuration
python-dotenv==1.0.0

# Web Framework
Flask==3.0.0
flask-cors==4.0.0
Flask-Limiter==3.5.0

# Utilities
tqdm==4.66.1
colorama==0.4.6

# Knowledge Graph (Third Brain)
networkx>=3.0

# Machine Learning & Embeddings
scikit-learn==1.3.2
numpy==1.26.2
tabulate
